[XLA:GPU] Use rendezvous to exchange pointers and CUDA events in memcpy ragged-all-to-all.
The previous implementation used NCCL AllGather to exchange target device pointers and synchronize streams at the start of the kernel, but it didn't synchronize streams at the end of the kernel. Since we use push model for memcpy, it could be that one stream progresses before all updates have arrived.

PiperOrigin-RevId: 732933504