[XLA] Enforce that instructions do not reference computations in other modules.
Fix a call to `HloComputation::set_parent()` in the `HloInstruction` cloning code that at least ephemerally creating a cross-module edge. As far as I can see this `set_parent()` call is completely superfluous, since the cloned instruction has not yet been added to a computation and hence does not have a parent.

We also make a small refactoring so called computations are set more consistently via a `HloInstruction::set_called_computation()` helper.

PiperOrigin-RevId: 733319330